Fix: Trying to stop a not running Tomcat interrupted the build process on Windows
Thus we have to ignore a return value of 1 after trying to stop the server. For this we
need to upgrade exec-maven-plugin to 1.6.0 in order to use option 'successCode'.

Furthermore, Tomcat under Windows needs run asynchronously in order for the
build to continue and actually execute the test. For this we use ''async=true'.