Bump Liferay Hot Deploy JMX Listener to 2.0.0-SNAPSHOT
It is ugly to use a snapshot (repository), but at least this version seems
to be compatible with LR 7.1 and does not need an additional copy of
com.liferay.portal:com.liferay.portal.kernel:2.0.0 with an older version
number. Probably this is what kep the installed modules from being
auto-activated in the first place.